netlify/functions/requirements.txt
ultralytics
opencv-python-headless
pillow
torch
torchvision
gdown
numpy

